+++ src/pages/admin/LibraryManage/Tab/BorrowingStatusTab.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const BorrowingStatusTab = () => {
+  return <>대출 현황</>;
+};
+
+export default BorrowingStatusTab;
