+++ src/pages/admin/LibraryManage/Tab/RequestManageTab.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const RequestManageTab = () => {
+  return <>승인 요청</>;
+};
+
+export default RequestManageTab;
